requirements.txt
python-dotenv
google-generativeai
transformers
sentencepiece
sentence-transformers
numpy
asyncio
streamlit
torch